next iteration of porting with tests etc

+++ constants.cpp
@@ -1,12 +1,12 @@
 /**
- * 
+ * constants.cpp
  */
 
 #include "constants.hpp"
 
-const float cuben::constants::iterTol = 1e-8;
-const float cuben::constants::zeroTol = 1e-8;
-const float cuben::constants::adaptiveTol = 1e-4;
-const float cuben::constants::relDiffEqTol = 1e-6;
-const float cuben::constants::bvpZeroTol = 1e-6;
-const int cuben::constants::iterLimit = 1e4;
+const float cuben::constants::iterTol = (float)1e-8;
+const float cuben::constants::zeroTol = (float)1e-8;
+const float cuben::constants::adaptiveTol = (float)1e-4;
+const float cuben::constants::relDiffEqTol = (float)1e-6;
+const float cuben::constants::bvpZeroTol = (float)1e-6;
+const int cuben::constants::iterLimit = (int)1e4;
